model_foundry_v2/requirements-v2.txt
Flask>=3.0,<4.0
Jinja2>=3.1,<4.0
Werkzeug>=3.0,<4.0
itsdangerous>=2.1,<3.0
click>=8.1,<9.0
numpy>=1.24,<2.0
Pillow>=10.0,<11.0
tensorflow==2.13.0
